[ci] Introduce a two-step release process #4700

+++ .ci/build.sh
@@ -106,7 +106,8 @@ function build() {
         pmd_ci_build_and_upload_doc
     pmd_ci_log_group_end
 
-    if pmd_ci_maven_isReleaseBuild; then
+    # release is published only for the case b) pmd-cli/pmd-dist release
+    if pmd_ci_maven_isReleaseBuild && [[ "${PMD_CI_TAG}" == *-dist ]]; then
     pmd_ci_log_group_start "Publishing Release"
         pmd_ci_gh_releases_publishRelease "$GH_RELEASE"
         pmd_ci_sourceforge_selectDefault "${PMD_CI_MAVEN_PROJECT_VERSION}"
@@ -114,10 +115,14 @@ function build() {
     pmd_ci_log_group_end
     fi
 
+    # create a baseline for snapshot builds (when pmd-dist is built)
+    # or for release builds for case b) when pmd-cli/pmd-dist is released
+    if pmd_ci_maven_isSnapshotBuild || [[ "${PMD_CI_TAG}" == *-dist ]]; then
     pmd_ci_log_group_start "Creating new baseline for regression tester"
         regression_tester_setup_ci
         regression_tester_uploadBaseline
     pmd_ci_log_group_end
+    fi
 
     #
     # everything from here runs only on snapshots, not on release builds
@@ -182,25 +187,51 @@ function pmd_ci_build_run() {
         pmd_ci_log_info "This is a snapshot build"
     fi
 
-    ./mvnw clean deploy -P${mvn_profiles} --show-version --errors --batch-mode "${PMD_MAVEN_EXTRA_OPTS[@]}"
+    # There are two possible builds:
+    if [[ "${PMD_CI_TAG}" != *-dist ]]; then
+        # a) pmd-core and languages modules
+        ./mvnw clean deploy -P${mvn_profiles},'!cli-dist' --show-version --errors --batch-mode "${PMD_MAVEN_EXTRA_OPTS[@]}"
+    else
+        # b) pmd-cli and pmd-dist
+        ./mvnw clean deploy -pl pmd-cli,pmd-dist --show-version --errors --batch-mode "${PMD_MAVEN_EXTRA_OPTS[@]}"
+    fi
 }
 
 #
 # Deploys the binary distribution
 #
 function pmd_ci_deploy_build_artifacts() {
-    # Deploy to sourceforge files https://sourceforge.net/projects/pmd/files/pmd/
-    pmd_ci_sourceforge_uploadFile "pmd/${PMD_CI_MAVEN_PROJECT_VERSION}" "pmd-dist/target/pmd-dist-${PMD_CI_MAVEN_PROJECT_VERSION}-bin.zip"
-    pmd_ci_sourceforge_uploadFile "pmd/${PMD_CI_MAVEN_PROJECT_VERSION}" "pmd-dist/target/pmd-dist-${PMD_CI_MAVEN_PROJECT_VERSION}-src.zip"
-    # Deploy SBOM
-    cp pmd-dist/target/bom.xml  "pmd-dist/target/pmd-${PMD_CI_MAVEN_PROJECT_VERSION}-cyclonedx.xml"
-    cp pmd-dist/target/bom.json "pmd-dist/target/pmd-${PMD_CI_MAVEN_PROJECT_VERSION}-cyclonedx.json"
-    pmd_ci_sourceforge_uploadFile "pmd/${PMD_CI_MAVEN_PROJECT_VERSION}" "pmd-dist/target/pmd-${PMD_CI_MAVEN_PROJECT_VERSION}-cyclonedx.xml"
-    pmd_ci_sourceforge_uploadFile "pmd/${PMD_CI_MAVEN_PROJECT_VERSION}" "pmd-dist/target/pmd-${PMD_CI_MAVEN_PROJECT_VERSION}-cyclonedx.json"
+    if pmd_ci_maven_isSnapshotBuild; then
+        # Deploy to sourceforge files https://sourceforge.net/projects/pmd/files/pmd/
+        pmd_ci_sourceforge_uploadFile "pmd/${PMD_CI_MAVEN_PROJECT_VERSION}" "pmd-dist/target/pmd-dist-${PMD_CI_MAVEN_PROJECT_VERSION}-bin.zip"
+        pmd_ci_sourceforge_uploadFile "pmd/${PMD_CI_MAVEN_PROJECT_VERSION}" "pmd-dist/target/pmd-dist-${PMD_CI_MAVEN_PROJECT_VERSION}-src.zip"
+        # Deploy SBOM
+        cp pmd-dist/target/bom.xml  "pmd-dist/target/pmd-${PMD_CI_MAVEN_PROJECT_VERSION}-cyclonedx.xml"
+        cp pmd-dist/target/bom.json "pmd-dist/target/pmd-${PMD_CI_MAVEN_PROJECT_VERSION}-cyclonedx.json"
+        pmd_ci_sourceforge_uploadFile "pmd/${PMD_CI_MAVEN_PROJECT_VERSION}" "pmd-dist/target/pmd-${PMD_CI_MAVEN_PROJECT_VERSION}-cyclonedx.xml"
+        pmd_ci_sourceforge_uploadFile "pmd/${PMD_CI_MAVEN_PROJECT_VERSION}" "pmd-dist/target/pmd-${PMD_CI_MAVEN_PROJECT_VERSION}-cyclonedx.json"
+    fi
 
-    if pmd_ci_maven_isReleaseBuild; then
+    # release build case a): only pmd-core and language modules released
+    if pmd_ci_maven_isReleaseBuild && [[ "${PMD_CI_TAG}" != *-dist ]]; then
         # create a draft github release
         pmd_ci_gh_releases_createDraftRelease "${PMD_CI_TAG}" "$(git rev-list -n 1 "${PMD_CI_TAG}")"
+        GH_RELEASE="$RESULT"
+    fi
+
+    # release build case b): pmd-cli and pmd-dist are released
+    if pmd_ci_maven_isReleaseBuild && [[ "${PMD_CI_TAG}" == *-dist ]]; then
+        # Deploy to sourceforge files https://sourceforge.net/projects/pmd/files/pmd/
+        pmd_ci_sourceforge_uploadFile "pmd/${PMD_CI_MAVEN_PROJECT_VERSION}" "pmd-dist/target/pmd-dist-${PMD_CI_MAVEN_PROJECT_VERSION}-bin.zip"
+        pmd_ci_sourceforge_uploadFile "pmd/${PMD_CI_MAVEN_PROJECT_VERSION}" "pmd-dist/target/pmd-dist-${PMD_CI_MAVEN_PROJECT_VERSION}-src.zip"
+        # Deploy SBOM
+        cp pmd-dist/target/bom.xml  "pmd-dist/target/pmd-${PMD_CI_MAVEN_PROJECT_VERSION}-cyclonedx.xml"
+        cp pmd-dist/target/bom.json "pmd-dist/target/pmd-${PMD_CI_MAVEN_PROJECT_VERSION}-cyclonedx.json"
+        pmd_ci_sourceforge_uploadFile "pmd/${PMD_CI_MAVEN_PROJECT_VERSION}" "pmd-dist/target/pmd-${PMD_CI_MAVEN_PROJECT_VERSION}-cyclonedx.xml"
+        pmd_ci_sourceforge_uploadFile "pmd/${PMD_CI_MAVEN_PROJECT_VERSION}" "pmd-dist/target/pmd-${PMD_CI_MAVEN_PROJECT_VERSION}-cyclonedx.json"
+
+        # draft release has already been created
+        pmd_ci_gh_releases_getLatestDraftRelease
         GH_RELEASE="$RESULT"
 
         # Deploy to github releases
@@ -221,7 +252,7 @@ function pmd_ci_build_and_upload_doc() {
     pmd_doc_create_archive
 
     pmd_ci_sourceforge_uploadFile "pmd/${PMD_CI_MAVEN_PROJECT_VERSION}" "docs/pmd-dist-${PMD_CI_MAVEN_PROJECT_VERSION}-doc.zip"
-    if pmd_ci_maven_isReleaseBuild; then
+    if pmd_ci_maven_isReleaseBuild && [[ "${PMD_CI_TAG}" != *-dist ]]; then
         pmd_ci_gh_releases_uploadAsset "$GH_RELEASE" "docs/pmd-dist-${PMD_CI_MAVEN_PROJECT_VERSION}-doc.zip"
     fi
 
@@ -230,19 +261,21 @@ function pmd_ci_build_and_upload_doc() {
     # Deploy javadoc to https://docs.pmd-code.org/apidocs/*/${PMD_CI_MAVEN_PROJECT_VERSION}/
     pmd_code_uploadJavadoc "${PMD_CI_MAVEN_PROJECT_VERSION}" "$(pwd)"
 
-    # render release notes
-    # updating github release text
-    rm -f .bundle/config
-    bundle config set --local path vendor/bundle
-    bundle config set --local with release_notes_preprocessing
-    bundle install
-    # renders, and skips the first 6 lines - the Jekyll front-matter
-    local rendered_release_notes
-    rendered_release_notes=$(bundle exec docs/render_release_notes.rb docs/pages/release_notes.md | tail -n +6)
-    local release_name
-    release_name="PMD ${PMD_CI_MAVEN_PROJECT_VERSION} ($(date -u +%d-%B-%Y))"
-    # Upload to https://sourceforge.net/projects/pmd/files/pmd/${PMD_CI_MAVEN_PROJECT_VERSION}/ReadMe.md
-    pmd_ci_sourceforge_uploadReleaseNotes "pmd/${PMD_CI_MAVEN_PROJECT_VERSION}" "${rendered_release_notes}"
+    if pmd_ci_maven_isSnapshotBuild || [[ "${PMD_CI_TAG}" != *-dist ]]; then
+        # render release notes
+        # updating github release text
+        rm -f .bundle/config
+        bundle config set --local path vendor/bundle
+        bundle config set --local with release_notes_preprocessing
+        bundle install
+        # renders, and skips the first 6 lines - the Jekyll front-matter
+        local rendered_release_notes
+        rendered_release_notes=$(bundle exec docs/render_release_notes.rb docs/pages/release_notes.md | tail -n +6)
+        local release_name
+        release_name="PMD ${PMD_CI_MAVEN_PROJECT_VERSION} ($(date -u +%d-%B-%Y))"
+        # Upload to https://sourceforge.net/projects/pmd/files/pmd/${PMD_CI_MAVEN_PROJECT_VERSION}/ReadMe.md
+        pmd_ci_sourceforge_uploadReleaseNotes "pmd/${PMD_CI_MAVEN_PROJECT_VERSION}" "${rendered_release_notes}"
+    fi
 
     if pmd_ci_maven_isSnapshotBuild && [ "${PMD_CI_BRANCH}" = "master" ]; then
         # only for snapshot builds from branch master: https://docs.pmd-code.org/snapshot -> pmd-doc-${PMD_CI_MAVEN_PROJECT_VERSION}
@@ -254,7 +287,7 @@ function pmd_ci_build_and_upload_doc() {
         pmd_ci_sourceforge_rsyncSnapshotDocumentation "${PMD_CI_MAVEN_PROJECT_VERSION}" "snapshot"
     fi
 
-    if pmd_ci_maven_isReleaseBuild; then
+    if pmd_ci_maven_isReleaseBuild && [[ "${PMD_CI_TAG}" != *-dist ]]; then
         # documentation is already uploaded to https://docs.pmd-code.org/pmd-doc-${PMD_CI_MAVEN_PROJECT_VERSION}
         # we only need to setup symlinks for the released version
         pmd_code_createSymlink "${PMD_CI_MAVEN_PROJECT_VERSION}" "latest"
